[refactor][style](ABC-622): Added class to separate horizontal and vertical logic and reordered functions alphabetically
